refactor: Update repository references - moved cabs_repository.go and its test file to the repositories folder
