+++ Backend/internal/repositories/cabs_repository.go
@@ -1,4 +1,4 @@
-package repository
+package repositories
 
 import (
 	"database/sql"
